requirements.txt
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
idna==2.7
jieba==0.39
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
Pillow==5.3.0
pycryptodome==3.7.2
pyparsing==2.3.0
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wordcloud==1.5.0
